+++ packages/safegen/src/anthropic/index.ts
@@ -1,1 +1,1 @@
-export * from "./anthropic-safegen-cached"
+export * from "./anthropic-safe-generator"
