+++ packages/safegen/src/anthropic/set-up-anthropic-json-generator.ts
