+++ packages/safegen/src/ollama/index.ts
@@ -1,1 +1,1 @@
-export * from "./ollama-safegen-cached"
+export * from "./ollama-safe-generator"
