+++ packages/safegen/src/ollama/ollama-safe-generator.ts
@@ -3,44 +3,41 @@ import { Squirrel } from "varmint"
 
 import type { GenerateFromSchema, SafeGenerator } from "../safegen"
 import { createSafeDataGenerator } from "../safegen"
-import type { ANTHROPIC_PRICING_FACTS } from "./anthropic-pricing-facts"
-import { buildAnthropicRequestParams } from "./build-anthropic-request-params"
-import type { GetUnknownJsonFromAnthropic } from "./set-up-anthropic-generator"
-import { setUpAnthropicJsonGenerator } from "./set-up-anthropic-generator"
+import { buildOllamaRequestParams } from "./build-ollama-request-params"
+import type { GetUnknownJsonFromOllama } from "./set-up-ollama-json-generator"
+import { setUpOllamaJsonGenerator } from "./set-up-ollama-json-generator"
 
-export type AnthropicSafeGenOptions = {
-	model: keyof typeof ANTHROPIC_PRICING_FACTS
+export type OllamaSafeGenOptions = {
+	model: `llama3.2:1b` | `llama3.2` | (string & {})
 	usdBudget: number
 	usdMinimum: number
-	apiKey: string
 	cachingMode: CacheMode
 	cacheKey?: string
 	logger?: Pick<Console, `error` | `info` | `warn`>
 }
 
-export class AnthropicSafeGenerator implements SafeGenerator {
+export class OllamaSafeGenerator implements SafeGenerator {
 	public usdBudget: number
 	public usdMinimum: number
-	public getUnknownJsonFromAnthropic: GetUnknownJsonFromAnthropic
-	public getUnknownJsonFromAnthropicSquirreled: Squirreled<GetUnknownJsonFromAnthropic>
+	public getUnknownJsonFromOllama: GetUnknownJsonFromOllama
+	public getUnknownJsonFromOllamaSquirreled: Squirreled<GetUnknownJsonFromOllama>
 	public squirrel: Squirrel
 
 	public constructor({
 		model,
 		usdBudget,
 		usdMinimum,
-		apiKey,
 		cachingMode,
-		cacheKey = `anthropic-safegen`,
+		cacheKey = `ollama-safegen`,
 		logger,
-	}: AnthropicSafeGenOptions) {
+	}: OllamaSafeGenOptions) {
 		this.usdBudget = usdBudget
 		this.usdMinimum = usdMinimum
 		this.squirrel = new Squirrel(cachingMode)
-		this.getUnknownJsonFromAnthropic = setUpAnthropicJsonGenerator(apiKey)
-		this.getUnknownJsonFromAnthropicSquirreled = this.squirrel.add(
+		this.getUnknownJsonFromOllama = setUpOllamaJsonGenerator()
+		this.getUnknownJsonFromOllamaSquirreled = this.squirrel.add(
 			cacheKey,
-			this.getUnknownJsonFromAnthropic,
+			this.getUnknownJsonFromOllama,
 		)
 		this.from = createSafeDataGenerator(async (...params) => {
 			if (this.usdBudget < this.usdMinimum) {
@@ -48,14 +45,14 @@ export class AnthropicSafeGenerator implements SafeGenerator {
 				const fallback = params[1]
 				return fallback
 			}
-			const anthropicParams = buildAnthropicRequestParams(model, ...params)
+			const ollamaParams = buildOllamaRequestParams(model, ...params)
 			const instruction = params[0]
 			const previouslyFailedResponses = params[3]
-			const response = await this.getUnknownJsonFromAnthropicSquirreled
+			const response = await this.getUnknownJsonFromOllamaSquirreled
 				.for(
 					`${instruction.replace(/[^a-zA-Z0-9-_. ]/g, `_`)}-${previouslyFailedResponses.length}`,
 				)
-				.get(anthropicParams)
+				.get(ollamaParams)
 			this.usdBudget -= response.usdPrice
 			return response.data
 		})
