+++ packages/safegen/src/ollama/set-up-ollama-json-generator.ts
