+++ packages/safegen/src/openai/set-up-openai-json-generator.ts
